requirements.txt
python-telegram-bot==21.0
flask==3.0.0
gunicorn==21.2.0